Check that raw pubkey data is formatted properly
We just check that the format is consistent (i.e uncompressed public keys have a length of 65 bytes and
start with 0x04), which is very fast, but not that the point is actually valid. This is much more expensive
and will be done if `checkValid` is true (which is the default).