Lazy initialization of PublicKey (#26)
* Delay pubkey conversion from raw to ec point

We just initialize public keys with raw data, and perform basic checks.
The actual conversion to an EC point will be done when it's needed.
This means that could create a pubkey with invalid data, we'll only find
out when we try to use it.

* make pubkey check optional at instantiation

This way it is safe by default, and we can choose to skip the
verification only when needed.

Also added a test.

* make test work with libsecp256k1

* Check that raw pubkey data is formatted properly

We just check that the format is consistent (i.e uncompressed public keys have a length 
of 65 bytes and start with 0x04), which is very fast, but not that the point is actually valid. 
This is much more expensive and will be done if `checkValid` is true (which is the default).