Crypto: Fix and extend PublicKey class
The workaround to skip pubkey validity checks was flawed because it
relied on an additional attribute that made the case class equality
fail.

PublicKey is now a regular class, public key data is checked by default
at creation time, and there is an optimized `toCompressedUnsafe`
method that allows fast (and unsafe) creation of public keys.