Unify scalar/private key and point/public key
We've removed scala and point types, and now have private keys which wrap a byte vector (32 bytes) and a compressed flag, and public keys which wrap a byte vector (33 or 65 bytes).
This is very similar to what is done in Bitcoin Core.
Keys are compressed by default, and we never create or use uncompressed public keys unless we explicitely need to.