Improve PSBT types
Add more structure to PSBT inputs and outputs.
The main change is to split segwit and non-segwit inputs, as they have
different requirements.

A higher-level application will always know whether the inputs it's
responsible for use segwit or not, so they'll be able to take advantage of
these new types (instead of manually parsing the PSBT fields themselves).